requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-mysql==1.5.2
python-dotenv==0.19.0
pyjwt==2.1.0